Allow only one trip to be updated from fuzzy matcher
