+++ src/ext/java/org/opentripplanner/ext/siri/TripAndPattern.java
@@ -1,0 +1,6 @@
+package org.opentripplanner.ext.siri;
+
+import org.opentripplanner.transit.model.network.TripPattern;
+import org.opentripplanner.transit.model.timetable.Trip;
+
+record TripAndPattern(Trip trip, TripPattern tripPattern) {}
